Renamed SingleShot connection file.
