+++ uitools/cpp/Esri/ArcGISRuntime/Toolkit/BasemapGalleryController.cpp
@@ -17,7 +17,7 @@
 
 // Toolkit headers
 #include "Internal/GeoViews.h"
-#include "Internal/SingleshotConnection.h"
+#include "Internal/SingleShotConnection.h"
 
 // ArcGISRuntime headers
 #include <Map.h>
