+++ uitools/cpp/Esri/ArcGISRuntime/Toolkit/Internal/SingleShotConnection.h
